UI Implementation
In order to show the rushing data and be able to manage it (sort &
filter), we have added a simple LiveView that, for now, is able to sort
by the requested fields and can search players using a search input.

Also added a new logo and updated the node modules.

Some logic changes have been added in order to finish the task.

This is a WIP.